+++ README.md
@@ -8,14 +8,14 @@ The scope of this codebase extends far beyond the scope of the paper linked abov
 
 ### Code structure
 
-In `src/`, the environments are defined in the `envs/` directory, and the approaches (both learning-based and not) are defined in the `approaches/` directory. The core NSRT learning algorithm happens in `src/nsrt_learning.py`, which has the following steps:
+In `src/`, the environments are defined in the `envs/` directory, and the approaches (both learning-based and not) are defined in the `approaches/` directory. The core NSRT learning algorithm happens in `src/nsrt_learning/nsrt_learning_main.py`, which has the following steps:
 * Segment data based on changes in predicates.
 * Learn how many NSRTs we need, along with the symbolic operator components of each (parameters, preconditions, and effects).
 * Learn options and annotate data with them.
 * Learn samplers.
 * Finalize the NSRTs.
 
-Methods for predicate learning are implemented as Approaches (e.g., `src/approaches/grammar_search_invention_approach.py`), and may interface with the core structure of `src/nsrt_learning.py` in various ways. Meanwhile, sampler learning and option learning are implemented in `src/sampler_learning.py` and `src/option_learning.py` respectively.
+Methods for predicate learning are implemented as Approaches (e.g., `src/approaches/grammar_search_invention_approach.py`), and may interface with the core structure of `src/nsrt_learning/nsrt_learning_main.py` in various ways. Meanwhile, sampler learning and option learning are implemented in `src/nsrt_learning/sampler_learning.py` and `src/nsrt_learning/option_learning.py` respectively.
 
 A simple implementation of search-then-sample task and motion planning is provided in `src/planning.py`. This implementation uses the "SeSamE" strategy: SEarch-and-SAMple planning, then Execution.
