+++ src/approaches/nsrt_learning_approach.py
@@ -10,7 +10,8 @@ from gym.spaces import Box
 from predicators.src.approaches import TAMPApproach
 from predicators.src.structs import Dataset, NSRT, ParameterizedOption, \
     Predicate, Type, Task, LowLevelTrajectory
-from predicators.src.nsrt_learning import learn_nsrts_from_data
+from predicators.src.nsrt_learning.nsrt_learning_main import \
+    learn_nsrts_from_data
 from predicators.src.settings import CFG
 from predicators.src import utils
