+++ tests/nsrt_learning/test_option_learning.py
@@ -6,12 +6,12 @@ from predicators.src.approaches import ApproachFailure
 from predicators.src.envs import create_env
 from predicators.src.ground_truth_nsrts import get_gt_nsrts
 from predicators.src.datasets.demo_replay import create_demo_replay_data
-from predicators.src.nsrt_learning import segment_trajectory, \
+from predicators.src.nsrt_learning.strips_learning import segment_trajectory, \
     learn_strips_operators
 from predicators.src.structs import STRIPSOperator
 from predicators.src.torch_models import MLPRegressor
-from predicators.src.option_learning import create_option_learner, \
-    _LearnedNeuralParameterizedOption
+from predicators.src.nsrt_learning.option_learning import \
+    create_option_learner, _LearnedNeuralParameterizedOption
 from predicators.src import utils
